producer/requirements.txt
kafka-python
requests
prometheus_client